feat: mnemonics allowed, refactor network config
